+++ doorstop/cli/test/test_all.py
@@ -124,33 +124,33 @@ class TestMain(unittest.TestCase):  # pylint: disable=R0904
 
 
 @unittest.skipUnless(os.getenv(ENV), REASON)  # pylint: disable=R0904
-class TestNew(TempTestCase):  # pylint: disable=R0904
+class TestCreate(TempTestCase):  # pylint: disable=R0904
 
-    """Integration tests for the 'doorstop new' command."""
+    """Integration tests for the 'doorstop create' command."""  # pylint: disable=C0103
 
-    def test_new(self):
-        """Verify 'doorstop new' can be called."""
-        self.assertIs(None, main(['new', '_TEMP', self.temp, '-p', 'REQ']))
+    def test_create(self):
+        """Verify 'doorstop create' can be called."""
+        self.assertIs(None, main(['create', '_TEMP', self.temp, '-p', 'REQ']))
 
-    def test_new_error_unknwon_parent(self):
-        """Verify 'doorstop new' returns an error with an unknown parent."""
+    def test_create_error_unknwon_parent(self):
+        """Verify 'doorstop create' returns an error with an unknown parent."""
         self.assertRaises(SystemExit, main,
-                          ['new', '_TEMP', self.temp, '-p', 'UNKNOWN'])
+                          ['create', '_TEMP', self.temp, '-p', 'UNKNOWN'])
 
-    def test_new_error_reserved_prefix(self):
-        """Verify 'doorstop new' returns an error with a reserved prefix."""
+    def test_create_error_reserved_prefix(self):
+        """Verify 'doorstop create' returns an error with a reserved prefix."""
         self.assertRaises(SystemExit, main,
-                          ['new', 'ALL', self.temp, '-p', 'REQ'])
+                          ['create', 'ALL', self.temp, '-p', 'REQ'])
 
 
-# @unittest.skipUnless(os.getenv(ENV), REASON)  # pylint: disable=R0904
+@unittest.skipUnless(os.getenv(ENV), REASON)  # pylint: disable=R0904
 class TestDelete(MockTestCase):  # pylint: disable=R0904
 
     """Integration tests for the 'doorstop delete' command."""
 
     def test_delete(self):
         """Verify 'doorstop delete' can be called."""
-        main(['new', 'PREFIX', 'prefix'])
+        main(['create', 'PREFIX', 'prefix'])
         self.assertIs(None, main(['delete', 'PREFIX']))
 
     def test_delete_error(self):
@@ -358,7 +358,7 @@ class TestImportFile(MockTestCase):  # pylint: disable=R0904
         """Verify 'doorstop import' can import a file using a custom map."""
         path = os.path.join(FILES, 'exported-map.csv')
         dirpath = os.path.join(self.temp, 'imported', 'prefix')
-        main(['new', 'PREFIX', dirpath])
+        main(['create', 'PREFIX', dirpath])
         # Act
         self.assertIs(None, main(['import', path, 'PREFIX',
                                   '--map', "{'mylevel': 'level'}"]))
@@ -379,7 +379,7 @@ class TestImportFile(MockTestCase):  # pylint: disable=R0904
         """Verify 'doorstop import' can import CSV to an existing document."""
         path = os.path.join(FILES, 'exported.csv')
         dirpath = os.path.join(self.temp, 'imported', 'prefix')
-        main(['new', 'PREFIX', dirpath])
+        main(['create', 'PREFIX', dirpath])
         # Act
         self.assertIs(None, main(['import', path, 'PREFIX']))
         # Assert
@@ -390,7 +390,7 @@ class TestImportFile(MockTestCase):  # pylint: disable=R0904
         """Verify 'doorstop import' can import TSV to an existing document."""
         path = os.path.join(FILES, 'exported.tsv')
         dirpath = os.path.join(self.temp, 'imported', 'prefix')
-        main(['new', 'PREFIX', dirpath])
+        main(['create', 'PREFIX', dirpath])
         # Act
         self.assertIs(None, main(['import', path, 'PREFIX']))
         # Assert
@@ -401,7 +401,7 @@ class TestImportFile(MockTestCase):  # pylint: disable=R0904
         """Verify 'doorstop import' can import XLSX to an existing document."""
         path = os.path.join(FILES, 'exported.xlsx')
         dirpath = os.path.join(self.temp, 'imported', 'prefix')
-        main(['new', 'PREFIX', dirpath])
+        main(['create', 'PREFIX', dirpath])
         # Act
         self.assertIs(None, main(['import', path, 'PREFIX']))
         # Assert