+++ CHANGES.md
@@ -12,6 +12,7 @@ Changelog
 - Made 'all' a reserved word that cannot be used as a prefix.
 - Added `doorstop import FILE PREFIX` to import from the export formats.
 - Added `doorstop deleete PREFIX` to delete document directories.
+- Renamed `doorstop new ...` to `doorstop create ...`.
 
 0.6 (2014/5/15)
 ---------------